+++ frontend/pages/ManageControlsPage/MacOSSettings/AggregateMacSettingsIndicators/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AggregateMacSettingsIndicators";
