+++ frontend/pages/ManageControlsPage/MacOSSettings/_styles.scss
@@ -15,37 +15,3 @@
     }
   }
 }
-
-.aggregate-mac-settings-indicators {
-  display: flex;
-  height: 94px;
-  border-top: 1px solid #e2e4ea;
-  border-bottom: 1px solid #e2e4ea;
-  border-left: 1px solid #e2e4ea;
-  border-radius: 6px;
-
-  .aggregate-mac-settings-indicator {
-    flex-grow: 1;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    border-right: 1px solid #e2e4ea;
-
-    // but not the tooltip text
-    .tooltip {
-      font-weight: $bold;
-      flex-direction: column;
-    }
-
-    a {
-      font-weight: $regular;
-    }
-  }
-
-  .aggregate-mac-settings-indicator:last-child {
-    border-top-right-radius: 6px;
-    border-bottom-right-radius: 6px;
-  }
-}
